Add harvest source types caching
We run into issues with behat with wrong harvest source types returned
by "dkan_harvest_source_types_definition" on a @AfterScenario call. The
possible cause being the Behat/Dkanextension context failing to properly
support drupal hooks implementation and calls. A possible solution for
this is to cache the harvest source types entries and bypass the hooks
invocation loop alltogether.

Included in this commit are:
* Caching code added to "dkan_harvest_source_types_definition".
* Enabling/Disable modules that implements the "harvest_source_types"
hook awearness.
* Phpunit tests.